+++ src/novelystlib/__init__.py
@@ -1,14 +1,17 @@
 """Package for yWriter tree view.
 
 Modules:
-yw7_work_file -- provide a class for yWriter 7 project editing.
-nv_exporter -- provide a converter class for yWriter export.
-tree_viewer -- provide a tkinter tree view class for novelyst.
-novelyst_tk -- provide a tkinter GUI framework for novelyst.
-element_view -- provide a generic class for viewing tree element properties.
-project_view -- provide a class for viewing and editing project properties.
 chapter_view -- provide a class for viewing and editing chapter properties.
+character_view -- provide a class for viewing and editing character properties.
+element_view -- provide a generic class for viewing tree element properties.
+label_combo -- provide a tkinter combobox with a label.
+label_entry -- provide a tkinter entry box with a label.
+novelyst_tk -- provide a tkinter GUI framework for novelyst.
+nv_exporter -- provide a converter class for yWriter export.
+project_view -- provide a class for viewing and editing project properties.
 scene_view -- provide a class for viewing and editing scene properties.
+tree_viewer -- provide a tkinter tree view class for novelyst.
+yw7_work_file -- provide a class for yWriter 7 project editing.
 
 Copyright (c) 2022 Peter Triesberger
 For further information see https://github.com/peter88213/novelyst