+++ src/novelystlib/project_view.py
@@ -21,45 +21,45 @@ class ProjectView(ElementView):
         """
         super(). __init__(ui, element)
         self._renChapters = tk.BooleanVar(value=element.kwVar['Field_RenumberChapters'])
-        self._renChaptersCheckbox = ttk.Checkbutton(ui._valuesWindow, text='Auto number chapters when refreshing the tree',
+        self._renChaptersCheckbox = ttk.Checkbutton(self._valuesFrame, text='Auto number chapters when refreshing the tree',
                                          variable=self._renChapters, onvalue=True, offvalue=False)
-        self._renChaptersCheckbox.pack(anchor='w', padx=5, pady=2)
+        self._renChaptersCheckbox.pack(anchor='w', pady=2)
 
         self._chHdPrefix = tk.StringVar(value=element.kwVar['Field_ChapterHeadingPrefix'])
-        self._chHdPrefixEntry = LabelEntry(ui._valuesWindow, text='Chapter heading prefix', textvariable=self._chHdPrefix)
-        self._chHdPrefixEntry.pack(anchor='w', padx=5, pady=2)
+        self._chHdPrefixEntry = LabelEntry(self._valuesFrame, text='Chapter heading prefix', textvariable=self._chHdPrefix)
+        self._chHdPrefixEntry.pack(anchor='w', pady=2)
 
         self._chHdSuffix = tk.StringVar(value=element.kwVar['Field_ChapterHeadingSuffix'])
-        self._chHdSuffixEntry = LabelEntry(ui._valuesWindow, text='Chapter heading suffix', textvariable=self._chHdSuffix)
-        self._chHdSuffixEntry.pack(anchor='w', padx=5, pady=2)
+        self._chHdSuffixEntry = LabelEntry(self._valuesFrame, text='Chapter heading suffix', textvariable=self._chHdSuffix)
+        self._chHdSuffixEntry.pack(anchor='w', pady=2)
 
         self._romanChapters = tk.BooleanVar(value=element.kwVar['Field_RomanChapterNumbers'])
-        self._romanChaptersCheckbox = ttk.Checkbutton(ui._valuesWindow, text='Use Roman chapter numbers',
+        self._romanChaptersCheckbox = ttk.Checkbutton(self._valuesFrame, text='Use Roman chapter numbers',
                                         variable=self._romanChapters, onvalue=True, offvalue=False)
-        self._romanChaptersCheckbox.pack(anchor='w', padx=5, pady=2)
+        self._romanChaptersCheckbox.pack(anchor='w', pady=2)
 
         self._renWithinParts = tk.BooleanVar(value=element.kwVar['Field_RenumberWithinParts'])
-        self._renWithinPartsCheckbox = ttk.Checkbutton(ui._valuesWindow, text='Reset chapter number when starting a new part',
+        self._renWithinPartsCheckbox = ttk.Checkbutton(self._valuesFrame, text='Reset chapter number when starting a new part',
                                         variable=self._renWithinParts, onvalue=True, offvalue=False)
-        self._renWithinPartsCheckbox.pack(anchor='w', padx=5, pady=2)
+        self._renWithinPartsCheckbox.pack(anchor='w', pady=2)
 
         self._renParts = tk.BooleanVar(value=element.kwVar['Field_RenumberParts'])
-        self._renPartsCheckbox = ttk.Checkbutton(ui._valuesWindow, text='Auto number parts when refreshing the tree',
+        self._renPartsCheckbox = ttk.Checkbutton(self._valuesFrame, text='Auto number parts when refreshing the tree',
                                         variable=self._renParts, onvalue=True, offvalue=False)
-        self._renPartsCheckbox.pack(anchor='w', padx=5, pady=2)
+        self._renPartsCheckbox.pack(anchor='w', pady=2)
 
         self._ptHdPrefix = tk.StringVar(value=element.kwVar['Field_PartHeadingPrefix'])
-        self._ptHdPrefixEntry = LabelEntry(ui._valuesWindow, text='Part heading prefix', textvariable=self._ptHdPrefix)
-        self._ptHdPrefixEntry.pack(anchor='w', padx=5, pady=2)
+        self._ptHdPrefixEntry = LabelEntry(self._valuesFrame, text='Part heading prefix', textvariable=self._ptHdPrefix)
+        self._ptHdPrefixEntry.pack(anchor='w', pady=2)
 
         self._ptHdSuffix = tk.StringVar(value=element.kwVar['Field_PartHeadingSuffix'])
-        self._ptHdSuffixEntry = LabelEntry(ui._valuesWindow, text='Part heading suffix', textvariable=self._ptHdSuffix)
-        self._ptHdSuffixEntry.pack(anchor='w', padx=5, pady=2)
+        self._ptHdSuffixEntry = LabelEntry(self._valuesFrame, text='Part heading suffix', textvariable=self._ptHdSuffix)
+        self._ptHdSuffixEntry.pack(anchor='w', pady=2)
 
         self._romanParts = tk.BooleanVar(value=element.kwVar['Field_RomanPartNumbers'])
-        self._romanPartsCheckbox = ttk.Checkbutton(ui._valuesWindow, text='Use Roman part numbers',
+        self._romanPartsCheckbox = ttk.Checkbutton(self._valuesFrame, text='Use Roman part numbers',
                                         variable=self._romanParts, onvalue=True, offvalue=False)
-        self._romanPartsCheckbox.pack(anchor='w', padx=5, pady=2)
+        self._romanPartsCheckbox.pack(anchor='w', pady=2)
 
     def apply_changes(self, ui):
         """Apply changes.
@@ -91,18 +91,3 @@ class ProjectView(ElementView):
         update_field(self._romanParts, 'Field_RomanPartNumbers')
         super().apply_changes(ui)
 
-    def close(self, ui):
-        """Remove widgets from the valuesWindow.
-        
-        Extends the superclass method.
-        """
-        super().close(ui)
-        self._renChaptersCheckbox.destroy()
-        self._chHdPrefixEntry.destroy()
-        self._chHdSuffixEntry.destroy()
-        self._romanChaptersCheckbox.destroy()
-        self._renWithinPartsCheckbox.destroy()
-        self._renPartsCheckbox.destroy()
-        self._ptHdPrefixEntry.destroy()
-        self._ptHdSuffixEntry.destroy()
-        self._romanPartsCheckbox.destroy()